Rename files to match projects Component
