+++ src/Components/App/App.tsx
@@ -2,7 +2,8 @@ import './App.css';
 import Header from '../Header/Header';
 import SideBar from '../SideBar/SideBar';
 import Landing from '../Landing/Landing';
-import Audio from '../Audio/Audio';
+import Audio from '../Projects/Projects';
+import Projects from '../Projects/Projects';
 import About from '../About/About';
 import Contact from '../Contact/Contact';
 import { Route, Routes } from 'react-router-dom';
@@ -16,7 +17,7 @@ const App = () => {
       <main>
         <Routes>
           <Route path="/" element={<Landing />} />
-          <Route path="/audio" element={<Audio />} />
+          <Route path="/projects" element={<Projects />} />
           <Route path="/about" element={<About />} />
           <Route path="/contact" element={<Contact />} />
         </Routes>
